requirements.txt
accelerate==1.8.1
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attr==0.3.2
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
brotlicffi==1.1.0.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
configparser==7.2.0
cryptography==45.0.5
Cython==3.1.2
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
docutils==0.21.2
et_xmlfile==2.0.0
executing==2.2.0
fanstatic==1.5
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.24.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
HTMLParser==0.0.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.2
hyperframe==6.1.0
idna==3.10
imagesize==1.4.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
js==1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
keyring==25.6.0
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langsmith==0.4.4
legacy-cgi==2.6.3
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
mpmath==1.3.0
narwhals==1.46.0
networkx==3.5
numpy==2.3.1
openai==1.93.1
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.3.1
parso==0.8.4
passlib==1.7.4
pillow==11.3.0
pipreqs==0.4.13
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
roman-numerals-py==3.1.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
streamlit==1.46.1
streamlit-tree-select==0.0.5
sympy==1.14.0
tenacity==8.5.0
thread==2.0.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
urllib3-secure-extra==0.1.0
watchdog==6.0.0
wcwidth==0.2.13
WebOb==1.8.9
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
yarg==0.1.10
zstandard==0.23.0
